Split TypeWidgetLibrary into interface and implementation
